Revert "Fixed static analysis warnings when built with x64."
This reverts commit a6e26c044315b8c8f439c26b4d4f50db604ef11f.

+++ ArrayAndStringManip/Matrix.h
@@ -35,7 +35,7 @@ namespace ArrayAndStringManip
 
 		int GetCellValue(int x, int y);
 
-		size_t GetDimension();
+		int GetDimension();
 
 	private:
 		void RotateCellsNinetyClockwise(int & topleft, int & bottomleft, int & bottomright, int & topright);
